Improve handling of generated files
Previously, every check was responsible for ignoring generated files,
and wasn't able to handle //line comments pointing from generated files
to input Go files, which is important for handling cgo.

Now, checks are registered with metadata (currently their ID and whether
errors in generated files should be ignored) and the lint package is
responsible for ignoring reported problems in generated files.

This causes an insubstantial amount of extra work because we check more
files now.